+++ src/lib/chunk.js
@@ -4,8 +4,8 @@
  */
 
 import { LocationQueue } from './util'
+import ndarray from 'ndarray'
 
-var ndarray = require('ndarray')
 
 export default Chunk
 
@@ -28,7 +28,8 @@ export default Chunk
  *
  */
 
-function Chunk(noa, requestID, ci, cj, ck, size, dataArray, userData) {
+/** @param {import('../index').Engine} noa */
+function Chunk(noa, requestID, ci, cj, ck, size, dataArray, userData, fillVoxelID = -1) {
     this.noa = noa
     this.isDisposed = false
 
@@ -53,8 +54,15 @@ function Chunk(noa, requestID, ci, cj, ck, size, dataArray, userData) {
     this._terrainMeshes = []
     noa._terrainMesher.initChunk(this)
     noa._objectMesher.initChunk(this)
+
     this._isFull = false
     this._isEmpty = false
+
+    this._wholeLayerVoxel = Array(size).fill(-1)
+    if (fillVoxelID >= 0) {
+        this.voxels.data.fill(fillVoxelID, 0, this.voxels.size)
+        this._wholeLayerVoxel.fill(fillVoxelID)
+    }
 
     // references to neighboring chunks, if they exist (filled in by `world`)
     var narr = Array.from(Array(27), () => null)
@@ -77,7 +85,7 @@ export function createVoxelArray(size) {
     return ndarray(arr, [size, size, size])
 }
 
-Chunk.prototype._updateVoxelArray = function (dataArray) {
+Chunk.prototype._updateVoxelArray = function (dataArray, fillVoxelID = -1) {
     // dispose current object blocks
     callAllBlockHandlers(this, 'onUnload')
     this.noa._objectMesher.disposeChunk(this)
@@ -88,6 +96,13 @@ Chunk.prototype._updateVoxelArray = function (dataArray) {
     this._blockHandlerLocs.empty()
     this.noa._objectMesher.initChunk(this)
     this.noa._terrainMesher.initChunk(this)
+
+    if (fillVoxelID >= 0) {
+        this._wholeLayerVoxel.fill(fillVoxelID)
+    } else {
+        this._wholeLayerVoxel.fill(-1)
+    }
+
     scanVoxelData(this)
 }
 
@@ -128,6 +143,11 @@ Chunk.prototype.set = function (i, j, k, newID) {
     var opaqueLookup = this.noa.registry._opacityLookup
     var handlerLookup = this.noa.registry._blockHandlerLookup
 
+    // track invariants about chunk data
+    if (!opaqueLookup[newID]) this._isFull = false
+    if (newID !== 0) this._isEmpty = false
+    if (this._wholeLayerVoxel[j] !== newID) this._wholeLayerVoxel[j] = -1
+
     // voxel lifecycle handling
     var hold = handlerLookup[oldID]
     var hnew = handlerLookup[newID]
@@ -146,17 +166,13 @@ Chunk.prototype.set = function (i, j, k, newID) {
     if (objOld) objMesher.setObjectBlock(this, 0, i, j, k)
     if (objNew) objMesher.setObjectBlock(this, newID, i, j, k)
 
-    // track full/emptiness and dirty flags for the chunk
-    if (!opaqueLookup[newID]) this._isFull = false
-    if (newID !== 0) this._isEmpty = false
-
+    // decide dirtiness states
     var solidityChanged = (solidLookup[oldID] !== solidLookup[newID])
     var opacityChanged = (opaqueLookup[oldID] !== opaqueLookup[newID])
+    var wasTerrain = !objOld && (oldID !== 0)
+    var nowTerrain = !objNew && (newID !== 0)
 
     if (objOld || objNew) this._objectsDirty = true
-
-    var wasTerrain = (!objOld && (oldID !== 0))
-    var nowTerrain = (!objNew && (newID !== 0))
     if (solidityChanged || opacityChanged || wasTerrain || nowTerrain) {
         this._terrainDirty = true
     }
@@ -231,31 +247,53 @@ Chunk.prototype.updateMeshes = function () {
 */
 
 function scanVoxelData(chunk) {
-    // flags for tracking if chunk is entirely opaque or transparent
-    var fullyOpaque = true
-    var fullyAir = true
-
-    chunk._blockHandlerLocs.empty()
     var voxels = chunk.voxels
     var data = voxels.data
     var len = voxels.shape[0]
     var opaqueLookup = chunk.noa.registry._opacityLookup
     var handlerLookup = chunk.noa.registry._blockHandlerLookup
     var objectLookup = chunk.noa.registry._objectLookup
+    var plainLookup = chunk.noa.registry._blockIsPlainLookup
     var objMesher = chunk.noa._objectMesher
-    for (var i = 0; i < len; ++i) {
-        for (var j = 0; j < len; ++j) {
+
+    // flags for tracking if chunk is entirely opaque or transparent
+    var fullyOpaque = true
+    var fullyAir = true
+
+    // scan vertically..
+    for (var j = 0; j < len; ++j) {
+
+        // fastest case where whole layer is air/dirt/etc
+        var layerID = chunk._wholeLayerVoxel[j]
+        if (layerID >= 0 && !objMesher[layerID] && !handlerLookup[layerID]) {
+            if (!opaqueLookup[layerID]) fullyOpaque = false
+            if (layerID !== 0) fullyAir = false
+            continue
+        }
+
+        var constantID = voxels.get(0, j, 0)
+
+        for (var i = 0; i < len; ++i) {
             var index = voxels.index(i, j, 0)
             for (var k = 0; k < len; ++k, ++index) {
                 var id = data[index]
-                // skip air blocks
+
+                // detect constant layer ID if there is one
+                if (constantID >= 0 && id !== constantID) constantID = -1
+
+                // most common cases: air block...
                 if (id === 0) {
                     fullyOpaque = false
                     continue
                 }
+                // ...or plain boring block (no mesh, handlers, etc)
+                if (plainLookup[id]) {
+                    fullyAir = false
+                    continue
+                }
+                // otherwise check opacity, object mesh, and handlers
                 fullyOpaque = fullyOpaque && opaqueLookup[id]
                 fullyAir = false
-                // handle object blocks and handlers
                 if (objectLookup[id]) {
                     objMesher.setObjectBlock(chunk, id, i, j, k)
                     chunk._objectsDirty = true
@@ -267,13 +305,14 @@ function scanVoxelData(chunk) {
                 }
             }
         }
+
+        if (constantID >= 0) chunk._wholeLayerVoxel[j] = constantID
     }
 
     chunk._isFull = fullyOpaque
     chunk._isEmpty = fullyAir
     chunk._terrainDirty = !chunk._isEmpty
 }
-